src/db/mongodb/account-repository/account.spec.ts
import { BcryptAdapter } from '../../../infra/criptography/bcrypt-adapter'
import { MongoHelper } from '../helpers/mongo-helper'
import { AccountMongoRepository } from './account'

describe('Account Mongo Repository', () => {
  beforeAll(async () => {
    await MongoHelper.connect(process.env.MONGO_URL)
  })

  afterAll(async () => {
    await MongoHelper.disconnect()
  })

  beforeEach(async () => {
    const account = MongoHelper.getCollection('accounts')
    await account.deleteMany({})
  })

  const makeSut = (): AccountMongoRepository => {
    return new AccountMongoRepository()
  }

  test('Should return an account on success', async () => {
    const sut = makeSut()

    const account = await sut.add({
      name: 'anyName',
      email: 'anyEmail@mail.com',
      password: 'anyPassword'
    })

    expect(account).toBeTruthy()
    expect(account.id).toBeTruthy()
    expect(account.name).toBe('anyName')
    expect(account.email).toBe('anyEmail@mail.com')
    expect(account.password).toBe('anyPassword')
  })
})
